+++ tenure/estimators/__init__.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from tenure.estimators.cox import CoxDiagnosticReport, CoxPH
 from tenure.estimators.kaplan_meier import KaplanMeier
+from tenure.estimators.logrank import LogRankReport, logrank_test
 from tenure.estimators.nelson_aalen import CumulativeHazardFunction, NelsonAalen
 from tenure.estimators.survival import SurvivalFunction
 from tenure.estimators.time_varying_cox import TimeVaryingCox
@@ -16,4 +17,6 @@ __all__ = [
     "TimeVaryingCox",
     "SurvivalFunction",
     "CumulativeHazardFunction",
+    "logrank_test",
+    "LogRankReport",
 ]